# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Nexus Control API
  version: 1.0.0
  description: 单 Deployment Control 首版合同。/auth/v1 供 Nexus Web 同源访问；internal 路由只供受信下游服务调用。
servers:
  - url: /
paths:
...
  /auth/v1/login:
    post:
      operationId: webLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebLoginInput"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "401":
          $ref: "#/components/responses/Error"
  /auth/v1/logout:
    post:
      operationId: webLogout
      security:
        - CookieSession: []
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "403":
          $ref: "#/components/responses/Error"
  /auth/v1/setup:
    post:
      operationId: webSetup
      security:
        - SetupToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupOwnerInput"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "401":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /auth/v1/profile:
    patch:
      operationId: webUpdateProfile
      security:
        - CookieSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileInput"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /auth/v1/profile/password:
    post:
      operationId: webChangePassword
      security:
        - CookieSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /auth/v1/profile/password/receipt:
    get:
      operationId: webPasswordReceipt
      security:
        - CookieSession: []
      parameters:
        - $ref: "#/components/parameters/RequestID"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
...
  /auth/v1/directory/members:
    get:
      operationId: listMemberDirectory
      description: 返回当前登录用户所在 Organization 的 active 真人目录，不包含角色、状态或登录时间。
      security:
        - CookieSession: []
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "401":
          $ref: "#/components/responses/Error"
...
  /auth/v1/organization/invitations:
    get:
      operationId: listOrganizationInvitations
      security:
        - CookieSession: []
      responses:
        "200": {$ref: "#/components/responses/Success"}
        "403": {$ref: "#/components/responses/Error"}
    post:
      operationId: createOrganizationInvitation
      security:
        - CookieSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateOrganizationInvitationInput"}
      responses:
        "200": {$ref: "#/components/responses/Success"}
        "400": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
  /auth/v1/organization/invitations/{invitation_id}:
    delete:
      operationId: revokeOrganizationInvitation
      security:
        - CookieSession: []
      parameters:
        - name: invitation_id
          in: path
          required: true
          schema: {type: string}
      responses:
        "200": {$ref: "#/components/responses/Success"}
        "403": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
  /auth/v1/organization/invitations/{invitation_id}/record:
    delete:
      operationId: deleteOrganizationInvitation
      description: Delete accepted, revoked or expired invitation records only. Does not change organization membership. Owners may delete all records; admins may delete member-role records only.
      security:
        - CookieSession: []
      parameters:
        - name: invitation_id
          in: path
          required: true
          schema: {type: string}
      responses:
        "200": {$ref: "#/components/responses/Success"}
        "403": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
  /auth/v1/organization-invitations/{token}:
    get:
      operationId: previewOrganizationInvitation
      parameters:
        - {$ref: "#/components/parameters/InvitationToken"}
      responses:
        "200": {$ref: "#/components/responses/Success"}
        "404": {$ref: "#/components/responses/Error"}
  /auth/v1/organization-invitations/{token}/accept:
    post:
      operationId: acceptOrganizationInvitation
      description: 同源 Origin。未登录提交账号资料注册加入；已登录提交空对象并加入，必须当前无组织，不重设账号密码。邀请单次消费。
      parameters:
        - {$ref: "#/components/parameters/InvitationToken"}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - {$ref: "#/components/schemas/AcceptOrganizationInvitationInput"}
                - {type: object, maxProperties: 0}
      responses:
        "200": {$ref: "#/components/responses/Success"}
        "400": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
...
  /api/control/v1/setup/owner:
    post:
      operationId: setupOwner
      security:
        - SetupToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupOwnerInput"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "401":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    NodeCredential:
      type: http
      scheme: bearer
      description: 注册时由设备安全保存的独立随机凭据，不是浏览器 Session 或服务 token。
    ServiceToken:
      type: http
      scheme: bearer
    SetupToken:
      type: http
      scheme: bearer
    CookieSession:
      type: apiKey
      in: cookie
      name: nexus_session
  parameters:
    InvitationToken:
      name: token
      in: path
      required: true
      schema:
        type: string
        minLength: 64
        maxLength: 64
...
    RequestID:
      name: request_id
      in: query
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
  schemas:
...
    SetupOwnerInput:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
        display_name:
          type: string
        password:
          type: string
          minLength: 8
        deployment_name:
          type: string
          maxLength: 128
        organization_name:
          type: string
          maxLength: 128
    WebLoginInput:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        password:
          type: string
          minLength: 1
          maxLength: 1024
    UpdateProfileInput:
      type: object
      additionalProperties: false
      required: [avatar]
      properties:
        avatar:
          type: string
          maxLength: 255
...
    CreateOrganizationInvitationInput:
      type: object
      additionalProperties: false
      required: [role]
      properties:
        role:
          type: string
          enum: [admin, member]
        expires_in_hours:
          type: integer
          minimum: 1
          maximum: 720
          default: 168
    AcceptOrganizationInvitationInput:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: {type: string, minLength: 3, maxLength: 64}
        display_name: {type: string, maxLength: 128}
        password: {type: string, minLength: 8, maxLength: 1024}
...
    ChangePasswordInput:
      type: object
      additionalProperties: false
      required: [request_id, current_password, new_password]
      properties:
        request_id:
          type: string
          minLength: 8
          maxLength: 128
        current_password:
          type: string
        new_password:
          type: string
          minLength: 8
...
    Envelope:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        data: {}
        details:
          type: object
          additionalProperties: true
  responses:
    Success:
      description: 请求成功，业务载荷位于 data。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    Error:
      description: 请求失败。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"